fix several review remarks
